requirements.txt
moviepy==2.1.2
mutagen==1.47.0
protobuf==6.30.2
python-dotenv==1.1.0
